Cohen's standardized mean difference.

Cohensdp() computes the Cohen's d (noted $d_p$) and its confidence interval in either within-subject, between-subject design and single-group design. For the between-subject design, MBESS already has an implementation based on the "pivotal" method but the present method is faster, using the method based on the Lambda prime distribution (Lecoutre 2007) . See Hedges (1981); Cousineau (2022, submitted); Goulet-Pelletier and Cousineau (2018) .

Cohensdp(statistics, design, gamma )

Arguments

statistics
a list of pre-computed statistics. The statistics to provide depend on the design: - for "between": m1, m2 the means of the two groups, s1, s2 the standard deviation of the two groups, and n1, n2, the sample sizes of the two groups; - for "within": m1, m2, s1, s2, n, and r or rho the correlation between the measure; - for "single": m, s, n and m0 the reference mean from which m is standardized).

design
the design of the measures ("within", "between", or "single");

gamma
the confidence level of the confidence interval (default 0.95)

Value

       the Cohen's $d_p$ statistic and its confidence interval.

Details

This function reduces the degrees of freedom by 1 in within-subject design when the population rho is unknown.

References

Cousineau D (2022). “The exact distribution of the Cohen's \(d_p\) in repeated-measure designs.” doi: 10.31234/osf.io/akcnd , https://psyarxiv.com/akcnd/.

Cousineau D (submitted). “The exact confidence interval of the Cohen's \(d_p\) in repeated-measure designs.” The Quantitative Methods for Psychology.

Goulet-Pelletier J, Cousineau D (2018). “A review of effect sizes and their confidence intervals, Part I: The Cohen's d family.” The Quantitative Methods for Psychology, 14(4), 242-265. doi: 10.20982/tqmp.14.4.p242 .

Hedges LV (1981). “Distribution theory for Glass's estimator of effect size and related estimators.” journal of Educational Statistics, 6(2), 107--128.

Lecoutre B (2007). “Another look at confidence intervals from the noncentral T distribution.” Journal of Modern Applied Statistical Methods, 6, 107 -- 116. doi: 10.22237/jmasm/1177992600 .

Examples

# example in which the means are 114 vs. 101 with sds of 14.3 and 12.5 respectively
Cohensdp( statistics = list( m1= 101, m2= 114, s1= 12.5, s2= 14.3, n1= 12, n2= 12 ), 
          design     = "between")
#> [1] -1.8074058 -0.9679684 -0.1090564

# example in a repeated-measure design
Cohensdp(statistics =list( m1= 101, m2= 114, s1= 12.5, s2= 14.3, n= 12, rho= 0.53 ),
         design     ="within")
#> [1] -1.5938740 -0.9679684 -0.3245819

# example with a single-group design where mu is a population parameter
Cohensdp( statistics = list( m = 101, m0 = 114, s = 12.5, n = 10 ), 
          design     = "single")
#> [1] -1.8019885 -1.0400000 -0.2424991

# The results can be displayed in three modes
res <- Cohensdp( statistics = list( m = 101, m0 = 114, s = 12.5, n = 10), 
                 design     = "single")

# a raw result of the Cohen's d_p and its confidence interval
res              
#> [1] -1.8019885 -1.0400000 -0.2424991

# a human-readable output
summarize( res ) 
#> Cohen's dp         = -1.040
#>   95.0% Confidence interval = [-1.802, -0.242]

# ... and a human-readable ouptut with additional explanations.
explain( res )   
#> Cohen's dp         = -1.040
#>  	 sample mean 101.000 is compared to assumed mean 114.000
#>  	 sample standard deviation 12.500 is the denominator
#>   95.0% Confidence interval = [-1.802, -0.242]
#>  	*: confidence interval obtained from the method lambda-prime with 9 degrees of freedom (Lecoutre, 2007, Journal of Modern Applied Statistical Methods)