The unbiased Hedges' standardized mean difference.

Hedgesgp() computes the unbiased Cohen's d (noted $g_p$) in either within-subject, between-subject design and single-group design. See h81,gc18;textualCohensdpLibrary.

Hedgesgp(statistics, design)

Arguments

statistics

a list of pre-computed statistics. The statistics to provide depend on the design: - for "between": m1, m2 the means of the two groups, s1, s2 the standard deviation of the two groups, and n1, n2, the sample sizes of the two groups; - for "within": m1, m2, s1, s2, n, and r or rho the correlation between the measure; - for "single": m, s, n and m0 the reference mean from which m is standardized).

design

the design of the measures ("within", "between", or "single");

Value

The unbiased Cohen's $d_p$ statistic, commonly called a Hedges' $g_p$. The return value is internally a dpObject which can be displayed with print, explain or summary/summarize.

Details

This function returns the Cohen's d_p statistics corrected for bias but no confidence interval as this estimate is not used to build such interval. This function uses r when rho is unknown.

References

Examples

# example in which the means are 114 vs. 101 with sDs of 14.3 and 12.5 respectively 
Hedgesgp( statistics = list( m1= 101, m2= 114, s1= 12.5, s2= 14.3, n1= 12, n2= 12 ), 
          design     = "between")
#> >>= CohensdpLibrary 0.5.11=>> There is no confidence interval for an (unbiased) Hedges's gp...
#> [1] -0.9345258

# example in a repeated-measure design
Hedgesgp( statistics = list( m1= 101, m2= 114, s1= 12.5, s2= 14.3, n= 12, rho= 0.53 ), 
          design     = "within")
#> >>= CohensdpLibrary 0.5.11=>> There is no confidence interval for an (unbiased) Hedges's gp...
#> [1] -0.9260117

# example with a single-group design where mu is a population parameter
Hedgesgp( statistics = list( m = 101, m0 = 114, s = 12.5, n = 10 ), 
          design     = "single")
#> >>= CohensdpLibrary 0.5.11=>> There is no confidence interval for an (unbiased) Hedges's gp...
#> [1] -0.9504299

# The results can be displayed in three modes
res <- Hedgesgp( statistics = list( m = 101, m0 = 114, s = 12.5, n = 10 ), 
                 design     = "single")
#> >>= CohensdpLibrary 0.5.11=>> There is no confidence interval for an (unbiased) Hedges's gp...

# a raw result of the Cohen's d_p and its confidence interval
res              
#> [1] -0.9504299

# a human-readable output
summarize( res ) 
#> Hedges' gp         = -0.950

# ... and a human-readable ouptut with additional explanations 
explain( res )   
#> Hedges' gp         = -0.950
#>  	 sample mean 101.000 is compared to assumed mean 114.000
#>  	 sample standard deviation 12.500 is the denominator