The K' or non-central K distribuiton.

The K' distribution was created to solve various problems in linear model statistics. pkprime returns the cumulative probability of the lambda prime distribution with parameters nu1, nu2, ncp; dkprime returns its density and qkprime, a quantile. Lecoutre (1999); Poitevineau and Lecoutre (2010) .

pkprime(x, nu1, nu2, ncp) 
dkprime(x, nu1, nu2, ncp) 
qkprime(p, nu1, nu2, ncp)

Arguments

x
the value from which a probability is sought;

nu1
the first degree of freedom;

nu2
the second degree of freedom;

ncp
the noncentrality parameter;

p
the probability from which a quantile is requested;

Value

the probability or quantile of a K' distribution

Details

kprime is a p,d,q set of functions that compute the K-prime distribution. This distribution has many applications, including to obtain the sampling distribution of _r_ given a population rho and the predictive distributions of rho given a sample _r_. See Lecoutre (1999); Poitevineau and Lecoutre (2010) .

These functions are herein implemented from the FORTRAN source code of Poitevineau and Lecoutre (2010) . Note that the library _sadists_ also implements this distribution (Pav 2020) . However, the sadists::kprime distribution is innacurate for small nu1 or small nu2.

References

Lecoutre B (1999). “Two useful distributions for Bayesian predictive procedures under normal models.” Journal of Statistical Planning and Inference, 79, 93 -- 105. doi: 10.1016/S0378-3758(98)00231-6 , https://doi.org/10.1016/S0378-3758(98)00231-6.

Pav SE (2020). “sadists: Some Additional Distributions [R package].” https://github.com/shabbychef/sadists.

Poitevineau J, Lecoutre B (2010). “Implementing Bayesian predictive procedures: The K-prime and K-square distributions.” Computational Statistics and Data Analysis, 54, 724 -- 731. doi: 10.1016/j.csda.2008.11.004 .

Poitevineau J, Lecoutre B (2010). “Statistical distributions for bayesian experimental data analysis fortran functions 1. continuous distributions.” https://eris62.eu.

Examples

dkprime(11.1, 9, 8, 10.0)  # 0.09410193
#> [1] 0.09410193
pkprime(11.1, 9, 8, 10.0)  # 0.606652
#> [1] 0.6066524
qkprime(0.01, 9, 8, 10.0)  # 3.875234
#> [1] 3.875234