hypergeometric functions.

The hypergeometric functions are a series of functions which includes the hypergeometric0F1, called the confluent hypergeometric limit function (D. Cousineau); the hypergeometric1F1, called the confluent hypergeometric function (Moreau 2014) ; and the hypergeometric2F1, called Gauss' confluent hypergeometric function (Michel and Stoitsov 2008) . These functions are involved in the computation of the K' and Lambda' distributions, as well as the Chi-square" and the t" distributions (Cousineau 2022) .

hypergeometric0F1(a, z)      
hypergeometric1F1(a, b, z)  
hypergeometric2F1(a, b, c, z)

Arguments

a

the first parameter;

z

the argument raised to the powers 0 ... infinity ;

b

the second parameter;

c

the third parameter;

Value

The result of the hypergeometric function.

References

Cousineau D (2022). “The exact distribution of the Cohen's \(d_p\) in repeated-measure designs.” doi:10.31234/osf.io/akcnd , https://psyarxiv.com/akcnd/.

Michel N, Stoitsov MV (2008). “Fast computation of the Gauss hypergeometric function with all its parameters complex with application to the Poschl-Teller-Ginocchio potential wave functions.” Computer Physics Communications, 178(7), 535-551. doi:10.1016/j.cpc.2007.11.007 .

Moreau J (2014). “Fortran Routines for Computation of Special Functions.” http://jean-pierre.moreau.pagesperso-orange.fr/fortran.html.

Examples

hypergeometric0F1(12, 0.4)         #   1.033851
#> [1] 1.033851
hypergeometric1F1(12, 14, 0.4)     #   1.409877
#> [1] 1.409877
hypergeometric2F1(12, 14, 16, 0.4) # 205.5699
#> [1] 205.5699