Lambda prime or noncentral Lambda distribution.

plprime computes the cumulative probability of the lambda-prime distribution with parameters nu, ncp. dlprime(x, nu, ncp) returns the density of the lambda prime and distribution qlprime(p, nu, ncp) its quantiles. See l99;textualCohensdpLibrary.

plprime(x, nu, ncp) 
dlprime(x, nu, ncp)
qlprime(p, nu, ncp)

Arguments

x

the score for which a probability is sought;

nu

the degree of freedom of the distribution;

ncp

the non-centrality parameter of the distribution;

p

the probability from which a quantile is requested;

Value

     The probability or quantile of a Lambda' distribution.

Details

lprime are functions that compute the Lambda-prime distribution. It was shown to be the predictive distribution of a population standardized mean or standardized mean difference in between-group design given an observed Cohen's dp l07CohensdpLibrary.

These functions are implemented from the FORTRAN source of pl10b;textualCohensdpLibrary. Note that the library sadists also implements this distribution sadists::lprime p20CohensdpLibrary.

References

Examples

dlprime(11.1, 9, 10.0) # 0.129447
#> [1] 0.129447
plprime(11.1, 9, 10.0) # 0.7134134
#> [1] 0.7134134
qlprime(0.01, 9, 10.0) # 4.2453
#> [1] 4.2453