Lambda second distribution or the non-central, non-standard, Lambda distribution.

This distribution was introduced in Cousineau (submitted) as the exact solution to the predictive distribution of the Cohen's dp in repeated-measure design. A more elegant notation was provided in Lecoutre (2022 - submitted) . It is the dual of the t" distribution, the sampling distribution of dp in repeated-measure design introduced in Cousineau (2022) .

plsecond(delta, n, d, rho) 
dlsecond(delta, n, d, rho) 
qlsecond(p,     n, d, rho)

Arguments

delta
the parameter of the population whose probability is to assess;

n
the sample size n

d
the observed d_p of the sample;

rho
the population correlation

p
the probability from which a quantile is requested

Value

     the probability or quantile of a Lambda'' distribution

Details

lsecond are p,d,q functions that compute the Lambda-second (L") distribution. This distribution is an generalization of the lambda-prime distribution (Lecoutre 1999) .

Note that the parameters are the raw sample size n, the observed Cohen's dp, and the population rho. All the scaling required are performed within the functions (and so you do not provide degrees of freedom).This is henceforth not a generic lambda-second distribution, but a lambda-second custom-tailored for the problem of standardized mean difference.

References

Cousineau D (2022). “The exact distribution of the Cohen's \(d_p\) in repeated-measure designs.” doi: 10.31234/osf.io/akcnd , https://psyarxiv.com/akcnd/.

Cousineau D (submitted). “The exact confidence interval of the Cohen's \(d_p\) in repeated-measure designs.” The Quantitative Methods for Psychology.

Lecoutre B (1999). “Two useful distributions for Bayesian predictive procedures under normal models.” Journal of Statistical Planning and Inference, 79, 93 -- 105. doi: 10.1016/S0378-3758(98)00231-6 .

Lecoutre B (2022 - submitted). “A note on the distributions of the sum and ratio of two correlated chi-square distributions.” submitted, submitted, submitted.

Examples

dlsecond(0.25, 9, 0.26, 0.333) # 1.03753
#> [1] 1.037531
plsecond(0.25, 9, 0.26, 0.333) # 0.494299
#> [1] 0.4942994
qlsecond(0.01, 9, 0.26, 0.333) # -0.6468003
#> [1] -0.6468