prior-informed Lambda second distribution.

This distribution extend the lambda second distribution introduced in Cousineau (submitted) as the exact solution to the predictive distribution of the Cohen's dp in repeated-measure when the population correlation is known. A more elegant notation was provided in Lecoutre (2022 - submitted) . The prior-informed lambda prime is a bayesian extension to the lambda prime distribution in the case where the population rho is not known. It is then replaced by a prior which indicates the probability of a certain rho given the observed correlation r.

ppilsecond(delta, n, d, r) 
dpilsecond(delta, n, d, r) 
qpilsecond(p,     n, d, r)

Arguments

delta

the parameter of the population whose probability is to assess;

n

the sample size n

d

the observed d_p of the sample;

r

the sample correlation

p

the probability from which a quantile is requested

Value

     The probability or quantile of a prior-informed Lambda'' distribution.

Details

pilsecond are p,d,q functions that compute the prior-informed Lambda-second (L") distribution. This distribution is an generalization of the lambda-prime distribution (Lecoutre 1999) . It can take up to two seconds to compute.

Note: the parameters are the raw sample size n, the observed Cohen's dp, and the sample correlation r. All the scaling required are performed within the functions (and so you do not provide degrees of freedom). This is henceforth not a generic lambda-second distribution, but a lambda-second custom-tailored for the problem of standardized mean difference.

References

Cousineau D (submitted). “The exact confidence interval of the Cohen's \(d_p\) in repeated-measure designs.” The Quantitative Methods for Psychology.

Lecoutre B (1999). “Two useful distributions for Bayesian predictive procedures under normal models.” Journal of Statistical Planning and Inference, 79, 93 -- 105. doi:10.1016/S0378-3758(98)00231-6 .

Lecoutre B (2022 - submitted). “A note on the distributions of the sum and ratio of two correlated chi-square distributions.” submitted, submitted, submitted.

Examples

### Note: this distribution can be slow to compute
### dpilsecond(0.25, 9, 0.26, 0.333) # 1.186735
### ppilsecond(0.25, 9, 0.26, 0.333) # 0.5150561
### qpilsecond(0.01, 9, 0.26, 0.333) # -0.7294266