The correction factor J for a standardized mean difference.

J() computes the correction factor to get an unbiased Cohen's $d_p$ in either within- subject, between-subject design and single-group design. See l22,gc18,;textualCohensdpLibrary.

J(statistics, design)

Arguments

statistics

a list of pre-computed statistics. The statistics to provide depend on the design: - for "between": n1, n2, the sample sizes of the two groups; - for "within": n, and r or rho the correlation between the measure; - for "single": n.

design

the design of the measures ("within", "between", or "single");

Value

The correction factor for unbiasing a Cohen's $d_p$. The return value is internally a dpObject which can be displayed with print, explain or summary/summarize.

Details

This function decreases the degrees of freedom by 1 in within-subject design when the population rho is unknown.

References

Examples

# example in which the means are 114 vs. 101 with sds of 14.3 and 12.5 respectively
J( statistics = list( n1 = 12, n2 = 12 ), 
   design     = "between")
#> [1] 0.9654507

# example in a repeated-measure design
J( statistics = list( n = 12, rho = 0.53 ), 
   design     = "within")
#> [1] 0.9566549

# example with a single-group design where mu is a population parameter
J( statistics = list( n = 12 ), 
   design     = "single")
#> [1] 0.9299598

# The results can be displayed in three modes
res <- J( statistics = list( n = 12 ), 
          design     = "single")

# a raw result of the Cohen's d_p and its confidence interval
res              
#> [1] 0.9299598

# a human-readable output
summarize( res ) 
#> Correction factor J(11) = 0.930

# ...and a human-readable ouptut with additional explanations
explain( res )   
#> Correction factor J(11) = 0.930
#>  	*: degrees of freedom nu = n-1 = 12-1; Hedges, 1981, Journal of Educational Statistics.