The K' or non-central K distribution.

The K' distribution was created to solve various problems in linear model statistics. pkprime returns the cumulative probability of the lambda prime distribution with parameters nu1, nu2, ncp; dkprime returns its density and qkprime, a quantile. l99,pl10;textualCohensdpLibrary.

pkprime(x, nu1, nu2, ncp) 
dkprime(x, nu1, nu2, ncp) 
qkprime(p, nu1, nu2, ncp)

Arguments

x

the value from which a probability is sought;

nu1

the first degree of freedom;

nu2

the second degree of freedom;

ncp

the noncentrality parameter;

p

the probability from which a quantile is requested;

Value

     The probability or quantile of a K' distribution.

Details

kprime is a (p,d,q) set of functions that compute the K-prime distribution. This distribution has many applications, including to obtain the sampling distribution of r given a population rho and the predictive distributions of rho given a sample r. See l99,pl10;textualCohensdpLibrary.

These functions are herein implemented from the FORTRAN source code of pl10b;textualCohensdpLibrary. Note that the library sadists also implements this distribution p20CohensdpLibrary. However, the sadists::kprime distribution is inaccurate for small nu1 or small nu2.

References

Examples

dkprime(11.1, 9, 8, 10.0)  # 0.09410193
#> [1] 0.09410193
pkprime(11.1, 9, 8, 10.0)  # 0.606652
#> [1] 0.6066524
qkprime(0.01, 9, 8, 10.0)  # 3.875234
#> [1] 3.875234