Lambda prime or noncentral Lambda distribution.

plprime computes the cumulative probability of the lambda-prime distribution with parameters nu, ncp. dlprime(x, nu, ncp) returns the density of the lambda prime and distribution qlprime(p, nu, ncp) its quantiles. See Lecoutre (1999) .

plprime(x, nu, ncp) 
dlprime(x, nu, ncp)
qlprime(p, nu, ncp)

Arguments

x
the score for which a probability is sought;

nu
the degree of freedom of the distribution;

ncp
the non-centrality parameter of the distribution;

p
the probability from which a quantile is requested;

Value

     the probability or quantile of a Lambda' distribution

Details

lprime are functions that compute the Lambda-prime distribution. It was shown to be the predictive distribution of a population standardized mean or standardized mean difference in between-group design given an observed Cohen's dp (Lecoutre 2007) .

These functions are implemented from the FORTRAN source of Poitevineau and Lecoutre (2010) . Note that the library sadists also implements this distribution sadists::lprime (Pav 2020) .

References

Lecoutre B (1999). “Two useful distributions for Bayesian predictive procedures under normal models.” Journal of Statistical Planning and Inference, 79, 93 -- 105. doi: 10.1016/S0378-3758(98)00231-6 .

Lecoutre B (2007). “Another look at confidence intervals from the noncentral T distribution.” Journal of Modern Applied Statistical Methods, 6, 107 -- 116. doi: 10.22237/jmasm/1177992600 .

Pav SE (2020). “sadists: Some Additional Distributions [R package].” https://github.com/shabbychef/sadists.

Poitevineau J, Lecoutre B (2010). “Statistical distributions for bayesian experimental data analysis fortran functions 1. continuous distributions.” https://eris62.eu.

Examples

dlprime(11.1, 9, 10.0) # 0.129447
#> [1] 0.129447
plprime(11.1, 9, 10.0) # 0.7134134
#> [1] 0.7134134
qlprime(0.01, 9, 10.0) # 4.2453
#> [1] 4.2453